refactor: package `td` -> package `testdata`
