+++ internal/testdata/channels.go
@@ -1,1 +1,1 @@
-package td
+package testdata
