+++ internal/testdata/doc.go
@@ -1,3 +1,3 @@
-// package td (TestData) provides literals and utility functions for generating objects-of-concern
+// package testdata (TestData) provides literals and utility functions for generating objects-of-concern
 // (objectives, channels, user-agents, etc) for other test packages.
-package td
+package testdata
